debug, planner, and new filters
Updated c.o.Debug to allow for printing field names and tuple values in intervals.

Changed planner to fail if traps are not contained within single Map or Reduce tasks. This prevents the chance of
multiple tasks writing to the same output location. Hadoop only partially supports appends, so it is not currently
possible to append subsequent jobs to existing trap files. Naming sections of a pipe assembly allows traps to be
bound to smaller sections of assemblies.

c.o.f.Sample and c.o.f.Limit Filters. Sample allows a given percentage of Tuples to pass. Limit only allows the
specified number of Tuples to pass.